LU-10499 pcc: keep PCC copy when it is being attached
When detach a file from PCC backend via FID, if the file is being
attached, it should not purge the coresponding PCC copy from the
PCC backend. Just keep the PCC copy to finish the attach process.

EX-3144 pcc: revalidate the pointer after attach
In this patch, we also fix a bug during PCC open attach. It
refreshes @pcci pointer since the lock may be released in
@pcc_try_readonly_open_attach().
Was-Change-Id: I470358dfde525e08e7110e862b30b527e5db94fe

EX-bug-id: EX-3133 EX-3144
Signed-off-by: Qian Yingjin <qian@ddn.com>
Change-Id: I8a8f7c6986d51eaf9b2516e5dd5a6f21aa38b7db
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/54389
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: James Simmons <jsimmons@infradead.org>
Reviewed-by: Li Xi <lixi@ddn.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>