LU-7372 tests: skip replay-dual test_24/25
Not sure which one of these subtests is causing problems, but
they are causing the following runtests test to hang unmounting
the MDS, just like test_26 was doing previously.

This is only a stopgap to confirm that one of these subtests is
causing the later unmount hang, and to get testing passing again.
There needs to be further isolation done to test_24 or test_25,
and to re-enable test_26, but that can be done afterward.

Test-Parameters: trivial testgroup=review-dne-part-2
Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I6d94eb040052b4912cf29ea37ca36ca4503ebbe5
Reviewed-on: https://review.whamcloud.com/43977
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: James Nunez <jnunez@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>