LUDOC-494 lnet: clarify use of route hopcount
Clarify that the default route hop setting is "undefined" (-1).
The default hopcount is interpreted as 1 when comparing routes
as part of the route selection procedure when sending.
The default hop setting may be used to create single-hop and
multi-hop routes.

Provide example to illustrate the case when hopcount should be set:
If avoid_asym_router_failure feature is enabled and single-hop
router NIs on the remote network fail to come up, using the default
hopcount will cause the route to be misinterpreted as multi-hop
and the router will still be selected for sending.

To avoid this case, set hop=1 if avoid_asym_router_failure
is enabled.

Signed-off-by: Serguei Smirnov <ssmirnov@whamcloud.com>
Change-Id: Ic0b97000b22389bb6950ad48070dd7d7682f9c56
Signed-off-by: Gian-Carlo DeFazio <defazio1@llnl.gov>
Reviewed-on: https://review.whamcloud.com/c/doc/manual/+/44916
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Olaf Faaland <faaland1@llnl.gov>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>